Fix bugs in handling of MaxRequests and RewriteLocation; add more tests.
MaxRequests: upon receiving too big a request, error 501 was returned,
which is not appropriate.  Change this to return 413 instead.  Introduced
new listener statement to customize the error message.

RewriteLocation: Minor bug in handling of the '1' value.  If location
pointed to the listener with the wrong protocol, it was not rewritten.

* config.c: New listener statement: Err413
* http.c (h413): New variable.
(do_http): Fix bug in max. request size handling: return 413,
instead of 510.
* pound.h (_listener): New field: err413.
* svc.c (need_rewrite): Fix listener protocol comparison.
* pound.8: Document Err413
* tests/Makefile.am: Add new tests.
* tests/testsuite.at: Likewise.
* tests/poundharness.pl: Fix cleanup after send.
Headers (both request and expected) may contain variables:
${LISTENER[N]} and ${BACKEND[N]}.
Header lines starting with dash mean that this header may not
be present in the request.
Fix cleanup after send and content comparison.
Add x-backend-ident header to the backend responses.
New endpoint: /redirect
* tests/config.at: Move definition of PT_CONF to testsuite.at
* tests/addheader.at: New test.
* tests/checkurl.at: New test.
* tests/errfile.at: New test.
* tests/headrem.at: New test.
* tests/maxrequest.at: New test.
* tests/rewriteloc.at: New test.
* tests/xhttp.at: Remove erroneous comment.