Clean-up in the service matching code.
This commit updates (and, partially, reverts) a816aada0f.
The HeaderRequest and HeaderDeny, introduced in a816aada0f,
are removed.  New statement "Header" is added as a replacement for
HeaderRequest.  The "HeadDeny" is implemented as "Not Header", which
should really be used instead of it.

* NEWS: Reflect changes.
* doc/pound.8: Likewise.
* README.md: Revise the docs.
* src/config.c: The HeaderRequest and HeaderDeny, introduced in a816aada0f,
are removed.  New statement "Header" is added as a replacement for
HeaderRequest.
* src/pound.h (BOOL_NOT): New constant.
(COND_HDR_DENY): Remove.  Use a combination of COND_BOOL(BOOL_NOT) and
COND_HDR to implement it.
(COND_HDR_REQ): Rename to COND_HDR.
(COND_COMPOUND): Rename to COND_BOOL.
* src/svc.c (match_cond): Update accordingly.