Fix watcher when running with supervisor.
The watcher must be set up after forking, in child process, otherwise
it will operate both in supervisor and server, which will create a race
condition (it is unclear which process will receive notifications).  On
the other hand, the set up must occur before switching to user privileges
and/or chrooting, so that these operations must be done in server as
well.

This improves commit 8c458cfa66.

* src/pound.c: Move call to watcher_setup from main to server.
Split chroot initialization in two parts: the preparatory part makes
sure the necessary file descriptors are open and libraries loaded.
This part is done in main. The chroot proper part and switching
to the user privileges is performed later, in server(), right
after setting up the watcher.